+++ src/main/java/org/example/module3/layered/repository/BrandRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface BrandRepository extends JpaRepository<BrandEntity,Long> {
+public interface BrandRepository extends JpaRepository<BrandEntity,Integer> {
 
     @Query("SELECT b FROM BrandEntity as b JOIN FETCH b.modelEntities")
     public List<BrandEntity> findAllOptimised();
